remchat_old/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-Session==0.4.0
Flask-Babel==2.0.0
Flask-SocketIO==5.3.6
python-socketio==5.8.0
python-engineio==4.8.0
eventlet==0.31.0
mysql-connector-python==8.0.26
Bcrypt
tinify